+++ providers/runtime.go
@@ -52,35 +52,6 @@ func (c *coordinator) RuntimeWithShutdownTimeout(timeout time.Duration) *Runtime
 	return runtime
 }
 
-func (c *coordinator) NewRuntime() *Runtime {
-	res := &Runtime{
-		coordinator: c,
-		providers:   map[string]*ConnectedProvider{},
-		schema: extensibleSchema{
-			loaded: map[string]struct{}{},
-			Schema: resources.Schema{
-				Resources: map[string]*resources.ResourceInfo{},
-			},
-		},
-		Recording:       NullRecording{},
-		shutdownTimeout: defaultShutdownTimeout,
-	}
-	res.schema.runtime = res
-
-	// TODO: do this dynamically in the future
-	res.schema.loadAllSchemas()
-	return res
-}
-
-func (c *coordinator) NewRuntimeFrom(parent *Runtime) *Runtime {
-	res := c.NewRuntime()
-	res.Recording = parent.Recording
-	for k, v := range parent.providers {
-		res.providers[k] = v
-	}
-	return res
-}
-
 type shutdownResult struct {
 	Response *plugin.ShutdownRes
 	Error    error
@@ -610,114 +581,9 @@ func (r *Runtime) AddSchema(name string, schema *resources.Schema) {
 	r.schema.Add(name, schema)
 }
 
-type extensibleSchema struct {
-	resources.Schema
-
-	loaded    map[string]struct{}
-	runtime   *Runtime
-	allLoaded bool
-	lockAll   sync.Mutex // only used in getting all schemas
-	lockAdd   sync.Mutex // only used when adding a schema
-}
-
-func (x *extensibleSchema) loadAllSchemas() {
-	x.lockAll.Lock()
-	defer x.lockAll.Unlock()
-
-	// If another goroutine started to load this before us, it will be locked until
-	// we complete to load everything and then it will be dumped into this
-	// position. At this point, if it has been loaded we can return safely, since
-	// we don't unlock until we are finished loading.
-	if x.allLoaded {
-		return
-	}
-	x.allLoaded = true
-
-	providers, err := ListActive()
-	if err != nil {
-		log.Error().Err(err).Msg("failed to list all providers, can't load additional schemas")
-		return
-	}
-
-	for name := range providers {
-		schema, err := x.runtime.coordinator.LoadSchema(name)
-		if err != nil {
-			log.Error().Err(err).Msg("load schema failed")
-		} else {
-			x.Add(name, schema)
-		}
-	}
-}
-
-func (x *extensibleSchema) Close() {
-	x.loaded = map[string]struct{}{}
-	x.Schema.Resources = nil
-}
-
-func (x *extensibleSchema) Lookup(name string) *resources.ResourceInfo {
-	if found, ok := x.Resources[name]; ok {
-		return found
-	}
-	if x.allLoaded {
+func (r *Runtime) asset() *inventory.Asset {
+	if r.Provider == nil || r.Provider.Connection == nil {
 		return nil
 	}
-
-	x.loadAllSchemas()
-	return x.Resources[name]
-}
-
-func (x *extensibleSchema) LookupField(resource string, field string) (*resources.ResourceInfo, *resources.Field) {
-	found, ok := x.Resources[resource]
-	if !ok {
-		if x.allLoaded {
-			return nil, nil
-		}
-
-		x.loadAllSchemas()
-
-		found, ok = x.Resources[resource]
-		if !ok {
-			return nil, nil
-		}
-		return found, found.Fields[field]
-	}
-
-	fieldObj, ok := found.Fields[field]
-	if ok {
-		return found, fieldObj
-	}
-	if x.allLoaded {
-		return found, nil
-	}
-
-	x.loadAllSchemas()
-	return found, found.Fields[field]
-}
-
-func (x *extensibleSchema) Add(name string, schema *resources.Schema) {
-	if schema == nil {
-		return
-	}
-	if name == "" {
-		log.Error().Msg("tried to add a schema with no name")
-		return
-	}
-
-	x.lockAdd.Lock()
-	defer x.lockAdd.Unlock()
-
-	if _, ok := x.loaded[name]; ok {
-		return
-	}
-
-	x.loaded[name] = struct{}{}
-	x.Schema.Add(schema)
-}
-
-func (x *extensibleSchema) AllResources() map[string]*resources.ResourceInfo {
-	if !x.allLoaded {
-		x.loadAllSchemas()
-	}
-
-	return x.Resources
+	return r.Provider.Connection.Asset
 }
